+++ BitcoinSPV/Sources/Core/WSKey.m
@@ -86,7 +86,8 @@ static NSData *WSKeyHMAC_DRBG(NSData *entropy, NSData *nonce);
 - (instancetype)initWithData:(NSData *)data compressed:(BOOL)compressed
 {
     if (data.length != WSKeyLength) {
-        DDLogVerbose(@"Incorrect private key data (length: %u != %u)", data.length, WSKeyLength);
+        DDLogVerbose(@"Incorrect private key data (length: %lu != %lu)",
+                     (unsigned long)data.length, (unsigned long)WSKeyLength);
         return nil;
     }
